Move utils/bundle.go to a new bundle package under application.
