+++ cmd/juju/application/bundlediff.go
@@ -23,6 +23,7 @@ import (
 	"github.com/juju/juju/api/modelconfig"
 	"github.com/juju/juju/apiserver/params"
 	jujucmd "github.com/juju/juju/cmd"
+	appbundle "github.com/juju/juju/cmd/juju/application/bundle"
 	"github.com/juju/juju/cmd/juju/application/utils"
 	"github.com/juju/juju/cmd/modelcmd"
 	"github.com/juju/juju/core/constraints"
@@ -129,7 +130,7 @@ func (c *bundleDiffCommand) Run(ctx *cmd.Context) error {
 		return errors.Trace(err)
 	}
 
-	bundle, err := utils.ComposeAndVerifyBundle(baseSrc, c.bundleOverlays)
+	bundle, err := appbundle.ComposeAndVerifyBundle(baseSrc, c.bundleOverlays)
 	if err != nil {
 		return errors.Trace(err)
 	}
@@ -221,7 +222,7 @@ func (c *bundleDiffCommand) bundleDataSource(ctx *cmd.Context) (charm.BundleData
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
-	bundleURL, _, err := utils.ResolveBundleURL(charmStore, c.bundle, c.channel)
+	bundleURL, _, err := appbundle.ResolveBundleURL(charmStore, c.bundle, c.channel)
 	if err != nil && !errors.IsNotValid(err) {
 		return nil, errors.Trace(err)
 	}
@@ -239,7 +240,7 @@ func (c *bundleDiffCommand) bundleDataSource(ctx *cmd.Context) (charm.BundleData
 		return nil, errors.Trace(err)
 	}
 
-	return utils.NewResolvedBundle(bundle), nil
+	return appbundle.NewResolvedBundle(bundle), nil
 }
 
 func (c *bundleDiffCommand) charmStore() (BundleResolver, error) {
@@ -268,7 +269,7 @@ func (c *bundleDiffCommand) readModel(apiRoot base.APICallCloser) (*bundlechange
 	if err != nil {
 		return nil, errors.Annotate(err, "getting model status")
 	}
-	model, err := utils.BuildModelRepresentation(status, c.makeModelExtractor(apiRoot), true, c.bundleMachines)
+	model, err := appbundle.BuildModelRepresentation(status, c.makeModelExtractor(apiRoot), true, c.bundleMachines)
 	return model, errors.Trace(err)
 }
 
@@ -284,7 +285,7 @@ func (c *bundleDiffCommand) getStatus(apiRoot base.APICallCloser) (*params.FullS
 	return &result, nil
 }
 
-func (c *bundleDiffCommand) makeModelExtractor(apiRoot base.APICallCloser) utils.ModelExtractor {
+func (c *bundleDiffCommand) makeModelExtractor(apiRoot base.APICallCloser) appbundle.ModelExtractor {
 	return &extractorImpl{
 		application: application.NewClient(apiRoot),
 		annotations: annotations.NewClient(apiRoot),
